+++ profile_management.ino
@@ -5,9 +5,9 @@ void clearProfiles()
 	memset(g_flowProfile, 0, sizeof(g_flowProfile)); // integer is more than 1 byte
 }
 
-void updateProfiles(int profileIndex, unsigned pumpPWM, long lastFlowPulseCount)
+void updateProfiles(uint16_t profileIndex, uint16_t pumpPWM, uint32_t lastFlowPulseCount)
 {
-	g_PWMProfile[profileIndex] = (byte) ((long) pumpPWM * 255 / 400); // store PWM setting 0-255 
+	g_PWMProfile[profileIndex] = (byte) ((uint32_t) pumpPWM * 255 / 400); // store PWM setting 0-255 
 	g_pressureProfile[profileIndex] = (byte)(g_averageP.mean() * 100.0 / 12.0); // store pressure average (at higher resolution of 0-100)
 	g_flowProfile[profileIndex] = (byte)(g_flowPulseCount - lastFlowPulseCount);
 }
